==> 4.clustering/6-analysing_clusters.Rmd <==
---
title: "6-Analysing clusters without doublets normalized"
author: 
- name: "Maria Lucia Romero Rivero"
  email: MLROMEROR@clinic.cat
date: "`r Sys.Date()`"
output:
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
    fig_captions: yes

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Pre-processing 

## Load packages


```{r message=FALSE, warning=FALSE}
library("RColorBrewer")
library(Signac)
library(Seurat)
library(GenomicRanges)
library(future)
#library(SeuratWrappers)
library(harmony)
library(EnsDb.Hsapiens.v86)
library(stringr)
library(dplyr)
library(ggplot2)
library(patchwork)
library(kableExtra)
library(tidyverse)
library(magick)
library(knitr) 

library(devtools)
library(reshape2)
library(plyr)
library(ComplexHeatmap)
library(binovisualfields)
library(viridis)
library(RColorBrewer)
set.seed(123)
```


## Parameters

```{r}
# Paths
path_to_obj <- here::here("~/Documents/multiome_tonsil_Lucia/results/R_objects/11.tonsil_multiome_integrated_without_doublets_normalized.rds")

path_to_markers<-here::here("~/Documents/multiome_tonsil_Lucia/results/tables/tonsil_markers_no_doublets_05.csv")

```


## Load data

```{r}
tonsil_wnn_without_doublet <- readRDS(path_to_obj)

tonsil_markers_02<-read_csv(path_to_markers)
```
 BCLL_14_T=T3"))
print(paste0( "There are ", sum(num_cell$initial_cells[3:4]), " in sample BCLL_15_T=T4"))
print(paste0( "There are ", sum(num_cell$initial_cells[5:7]), " in sample BCLL_2_T=T5"))
print(paste0( "There are ", sum(num_cell$initial_cells[8:9]), " in sample BCLL_8_T=T1"))
print(paste0( "There are ", sum(num_cell$initial_cells[10:11]), " in sample BCLL_9_T=T2"))

```{r}
# Save old identity classes (the cluster labels) for reference.
Idents(tonsil_wnn_without_doublet)<-"library_name"

# Rename classes.

tonsil_wnn_without_doublet <- RenameIdents(tonsil_wnn_without_doublet, 'BCLL_14_T_1'= "T3_a", 'BCLL_14_T_2'= "T3_b", 'BCLL_15_T_1'="T4_a", 'BCLL_15_T_2'="T4_b", 'BCLL_2_T_1'="T5_a",'BCLL_2_T_2'="T5_b", 'BCLL_2_T_3'="T5_c",'BCLL_8_T_1'="T1_a",'BCLL_8_T_2'="T1_b", 'BCLL_9_T_1'="T2_a", 'BCLL_9_T_2'="T2_b") 


tonsil_wnn_without_doublet[["new_library_name"]] <- Idents(object = tonsil_wnn_without_doublet)
DimPlot(tonsil_wnn_without_doublet, reduction = "wnn.umap",group.by = "library_name", pt.size = 0.5,cols = c('#a6cee3','#ffff99','#1f78b4','#b2df8a','#33a02c','#fb9a99','#e31a1c','#fdbf6f','#ff7f00','#6a3d9a','#cab2d6'))
```
```{r}
DimPlot(tonsil_wnn_without_doublet, reduction = "wnn.umap",group.by = "new_library_name", pt.size = 0.5,cols = c('#a6cee3','#ffff99','#1f78b4','#b2df8a','#33a02c','#fb9a99','#e31a1c','#fdbf6f','#ff7f00','#6a3d9a','#cab2d6'))
```

```{r fig.wide=TRUE}
vars <- str_subset(colnames(tonsil_wnn_without_doublet@meta.data), "^wsnn_res")
clusters_gg <- purrr::map(vars, function(x) {
  p <- DimPlot(
    tonsil_wnn_without_doublet,
    group.by = x,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T
  )
  p 
})
clusters_gg
```
```{r}
# "CD4 T-cells"="#1F78B4" 
# "CD8 T-cell"= "#33A02C" 
# "NBC/MBC"="#A6CEE3" 

# "GC"=#B2DF8A" 

# "PC"=#FB9A99" 
# "Non-lymphoid"="#E31A1C" 
# "Unclassified"= "#FDBF6F"

Idents(tonsil_wnn_without_doublet)<-"wsnn_res.0.05"

new.cluster.ids <- c("NBC/MBC", "CD4 T-cells","GCBC","GCBC","CD8 T-cell","PC","Non-lymphoid","Unclassified")
names(new.cluster.ids) <- levels(tonsil_wnn_without_doublet)
tonsil_wnn_without_doublet <- RenameIdents(tonsil_wnn_without_doublet, new.cluster.ids) 


new.level<-c("CD4 T-cells","CD8 T-cell","NBC/MBC","GCBC","PC","Non-lymphoid","Unclassified")
tonsil_wnn_without_doublet@active.ident <- factor(tonsil_wnn_without_doublet@active.ident, levels = new.level)
tonsil_wnn_without_doublet[["cluster_celltype_names"]] <- Idents(object = tonsil_wnn_without_doublet)

cluster_color<-setNames(c("#1F78B4", "#33A02C" ,"#A6CEE3", "#B2DF8A","#FB9A99" ,"#E31A1C", "#FDBF6F"),
                        levels(tonsil_wnn_without_doublet))

DimPlot(tonsil_wnn_without_doublet, reduction = "wnn.umap", pt.size = 0.5,cols = cluster_color)
```
## number of cell in each cluster

```{r}
cell<-as.data.frame(table(tonsil_wnn_without_doublet$cluster_celltype_names))
cell
```

```{r}
barp<-ggbarplot(cell,"Var1","Freq",fill  = "Var1",color = "Var1",palette = cluster_color,label = TRUE, lab.pos = "out", lab.col = "black",xlab = FALSE,ylab = "number of cells")
pdf(paste(figures_folder, "barplot_numbercell.pdf", sep=""),width=5,height=5,paper='special')


ggpar(barp,
      legend.title = "cell type")
dev.off()
```

### Get top n markers of each cluster


**Resolution 0.02**
 
```{r}
tonsil_markers_02 %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC) %>% write.csv(.,file=paste0("~/Documents/multiome_tonsil_Lucia/results/tables/", "top10_tonsil_markers_no_doublets_05.csv"))

tonsil_markers_02 %>% group_by(cluster) %>% top_n(n = 5, wt = avg_log2FC) %>% write.csv(.,file=paste0("~/Documents/multiome_tonsil_Lucia/results/tables/", "top5_tonsil_markers_no_doublets_05.csv"))
```
```{r}

top5_tonsil_markers_02<-tonsil_markers_02 %>% group_by(cluster) %>% top_n(n = 5, wt = avg_log2FC)
top7_tonsil_markers_02<-tonsil_markers_02 %>% group_by(cluster) %>% top_n(n = 7, wt = avg_log2FC)

top10_tonsil_markers_02<-tonsil_markers_02 %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC)
```

```{r}
df_top5<-as.data.frame(top5_tonsil_markers_02)
kbl(df_top5,caption = "Table of the top 5 marker of each cluster resolution 0.005") %>%
  kable_paper("striped", full_width = F)
```

```{r}
df_top7<-as.data.frame(top7_tonsil_markers_02)
df_mark<-as.data.frame(tonsil_markers_02)
kbl(df_top7,caption = "Table of the top 5 marker of each cluster resolution 0.005") %>%
  kable_paper("striped", full_width = F)
```


```{r}
#install.packages("htmlwidgets", type = "binary")
#install.packages("DT", type = "binary")

DT::datatable(df_top7)

```
```{r}
DT::datatable(df_mark)

```
```{r}
dot.10 <- DotPlot(tonsil_wnn_without_doublet, features = unique(top10_tonsil_markers_02$gene),cols = 'RdBu', cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) +ggtitle("res 0.05 top 10 of each cluster")

dot.5 <- DotPlot(tonsil_wnn_without_doublet, features = unique(top5_tonsil_markers_02$gene),cols = 'RdBu', cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) +ggtitle("res 0.05 top 5 of each cluster")

```

```{r}
dot.10 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))
dot.5 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 7))
```


## Complexheatmap

```{r}
m<-c("SUGCT","BANK1", "CXCR4", "MS4A1","RALGPS2","CD83","AICDA","BCL2A1","MEF2B", "BCL6","XBP1","IRF4","PRDM1","SLAMF7", "MZB1", "FKBP11","NKG7", "CD3D", "IL7R","LYZ", "S100A8")

```
```{r}
dot<-DotPlot(tonsil_wnn_without_doublet, features = m,cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) +ggtitle("res 0.005 markers of each cluster")
dot+
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 7))
```



```{r}
df<- dot$data
head(df)
```
```{r}
### the matrix for the scaled expression 
exp_mat<-df %>% 
  dplyr::select(-pct.exp, -avg.exp) %>%  
  pivot_wider(names_from = id, values_from = avg.exp.scaled) %>% 
  as.data.frame() 
    
row.names(exp_mat) <- exp_mat$features.plot  
exp_mat <- exp_mat[,-1] %>% as.matrix()

head(exp_mat)
```

```{r}
## the matrix for the percentage of cells express a gene

percent_mat<-df %>% 
  dplyr::select(-avg.exp, -avg.exp.scaled) %>%  
  pivot_wider(names_from = id, values_from = pct.exp) %>% 
  as.data.frame() 
    
row.names(percent_mat) <- percent_mat$features.plot  
percent_mat <- percent_mat[,-1] %>% as.matrix()

head(percent_mat)
```

```{r}
## the range is from 0 - 100
range(percent_mat)
```

```{r}
## these two matrix have the same dimension
dim(exp_mat)
```
```{r}
dim(percent_mat)
```

```{r}
colnames(exp_mat)
ncol(exp_mat)
```
```{r}

col_fun = circlize::colorRamp2(c(-1, 0, 2), viridis(20)[c(1,10, 20)])


cell_fun = function(j, i, x, y, w, h, fill){
          grid.rect(x = x, y = y, width = w, height = h, 
                    gp = gpar(col = NA, fill = NA))
          grid.circle(x=x,y=y,r= pindex(percent_mat, i, j)/100 * min(unit.c(w, h)),
                      gp = gpar(fill = col_fun(exp_mat[i, j]), col = NA))}
## also do a kmeans clustering for the genes with k = 4
cell_type<-c("CD4 T-cells","CD8 T-cell","NBC/MBC", "GCBC","PC","Non-lymphoid","Unclassified")
column_ha<- HeatmapAnnotation(
    cell_type = cell_type,
    col = list(cell_type = cluster_color),
    na_col = "grey"
)


Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 7),
       #row_km = 6,
        #border = "black",
        show_column_dend = FALSE,
        row_gap = unit(0, "mm"),
       show_column_names = FALSE,
       top_annotation = column_ha,
       cluster_columns = FALSE)
```

```{r}
Heatmap(exp_mat,
        heatmap_legend_param=list(title="exp"),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 5),
       #row_km = 6,
        #border = "black",
        show_column_dend = FALSE,
       #show_row_dend = FALSE,
        row_gap = unit(2, "mm"),
       top_annotation = column_ha,
       cluster_columns = FALSE)
```


```{r}
p <- Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 5),
       # row_km = 6,
  
        show_column_dend = FALSE,
        row_gap = unit(2, "mm"),
        cluster_columns = FALSE)
#rasterize(p, dpi = 300)
pdf(paste("/Users/mlromeror/Documents/multiome_tonsil_Lucia/Figures/
", "Dotplot_all_data.pdf", sep=""),width=5,height=5,paper="special")
p
dev.off()
```
```{r}


Heatmap(exp_mat,
        heatmap_legend_param=list(title="expression"),
        column_title = "clustered dotplot", 
        col=col_fun,
        rect_gp = gpar(type = "none"),
        cell_fun = cell_fun,
        row_names_gp = gpar(fontsize = 7),
       #row_km = 6,
        #border = "black",
        show_column_dend = FALSE,
        row_gap = unit(2, "mm"),
       show_column_names = FALSE,
       top_annotation = column_ha,
       cluster_columns = FALSE)


Heatmap(exp_mat, name = "expression", 
        
        row_names_gp = gpar(fontsize = 7),
        show_column_dend = FALSE,
        col=col_fun,
        column_split = factor(cell_type),
         top_annotation = column_ha,
        show_column_names = FALSE,
        cluster_columns = FALSE,
        show_row_dend = FALSE,
        )

row.subsections <- c(5,4,3,2,1,1,5)
row_split = data.frame(rep(cell_type, row.subsections))


pa = cluster::pam(exp_mat,k=7)


Heatmap(exp_mat, name = "expression", 
        
        row_names_gp = gpar(fontsize = 7),
        show_column_dend = FALSE,
        col=col_fun,
        column_split = factor(cell_type),
        row_split= pa$clustering,
        top_annotation = column_ha,
        #show_row_dend = FALSE,
        show_column_names = FALSE,
        cluster_columns = FALSE
        )
```


# Weights

```{r}
weight.vp<- VlnPlot(tonsil_wnn_without_doublet, features = "RNA.weight", group.by = "cluster_celltype_names",pt.size = 0, cols = cluster_color)+NoLegend()
p.weight.vp<- VlnPlot(tonsil_wnn_without_doublet, features = "peaks.weight", group.by = "cluster_celltype_names",pt.size = 0,cols = cluster_color)+NoLegend()

weight.vp + ggtitle("RNA modality weight")+theme(axis.text.x = element_text(angle = 20))+ NoLegend()
p.weight.vp+ ggtitle("ATAC modality weight")+theme(axis.text.x = element_text(angle =20))+ NoLegend()
```


# Top markers
```{r}
top7mark_cluster0<-top7_tonsil_markers_02[["gene"]][1:7]
top7mark_cluster1<-top7_tonsil_markers_02[["gene"]][8:14]
top7mark_cluster2<-top7_tonsil_markers_02[["gene"]][15:21]
top7mark_cluster3<-top7_tonsil_markers_02[["gene"]][22:28]
top7mark_cluster4<-top7_tonsil_markers_02[["gene"]][29:35]
top7mark_cluster5<-top7_tonsil_markers_02[["gene"]][36:42]
top7mark_cluster6<-top7_tonsil_markers_02[["gene"]][43:49]
top7mark_cluster7<-top7_tonsil_markers_02[["gene"]][50:56]

```


# Markers exploration
```{r}
markers_gg <- function(x){purrr::map(x, function(x) {
  
  p <- FeaturePlot(
    tonsil_wnn_without_doublet,
    features = x,
    reduction = "wnn.umap",
    pt.size = 0.1
  )
  p
})}
```


## Inaki markers 

```{r}
m<-c("PRDM1","XBP1","IRF4","MEF2B","BCL6")

DZ<-c("SUGCT", "CXCR4", "AICDA")

LZ<- c("CD83","BCL2A1")

GC<- c("MEF2B", "BCL6","IRF4")

PC<- c("PRDM1","SLAMF7", "MZB1", "FKBP11")

```


## cluster 0; memory and naive b cells 



```{r}
markers_gg(top7mark_cluster0)
```
```{r}
naive_markers<-c("CD79A", "CD79B", "BLNK")
memory_markers<-c("CD27")

```

```{r}
markers_gg(naive_markers)
```

```{r}
markers_gg(memory_markers)

```

```{r}
markers_gg(c("MS4A1","NT5E"))
```

## cluster 1: naive CD4 T-celL

IL6ST: naive CD4 T-cel

CCR7, CD62L, and CD45RA
```{r}
cd4<- c("CCR7")
```

```{r}
markers_gg(top7mark_cluster1)
```
```{r}
markers_gg(cd4)
```

## cluster 2: GC, 


```{r}
markers_gg(top7mark_cluster2)
```

```{r}
markers_gg(GC)

```


```{r}
markers_gg(LZ)
```

```{r}
markers_gg(DZ)

```

## cluster 3:  NK cd8+

ccl5:cd8t cell, nk

AOAH: NK
GNLY:NK


```{r}
markers_gg(top7mark_cluster3)
```


## cluster 4:PC


```{r}
markers_gg(top7mark_cluster4)
```

```{r}
markers_gg("KIR2DL4")
```
```{r}
markers_gg(PC)
```


## cluster 5:


```{r}
markers_gg(top7mark_cluster5)
```

```{r}
markers_gg( "MYO1E")
```


## cluster 6: 

```{r}
markers_gg(top7mark_cluster6)
```

```{r}
monocytes_markers<-c("LYZ","S100A8")
```

```{r}
markers_gg(monocytes_markers)
```

## cluster 7: doublets??

```{r}
markers_gg(top7mark_cluster7)
```

# Number of cell in each cluster


```{r}
DimPlot(tonsil_wnn_without_doublet, reduction = "wnn.umap", label = TRUE, pt.size = 0.5) 

```



```{r}
cell.num <- table(Idents(tonsil_wnn_without_doublet))
cell.num
```

# Rename clusters


```{r}
cell.num <- table(Idents(tonsil_wnn_without_doublet))
cell.num
```

```{r}
new.cluster.ids <- c("Naive/MBC", "Naive CD4 T-celL","GC/DZ", "GC/LZ", "NK T-cell", "PC", "Monocytes","NI")
names(new.cluster.ids) <- levels(tonsil_wnn_without_doublet)
tonsil_wnn_without_doublet <- RenameIdents(tonsil_wnn_without_doublet, new.cluster.ids)
DimPlot(tonsil_wnn_without_doublet, reduction = "wnn.umap", label = TRUE, pt.size = 0.5) 
```



# Bibliography  markers

**MARKERS**

Immature B cells express CD19, CD 20, CD34, CD38, and CD45R,
T-cell receptor/CD3 complex (TCR/CD3 complex) 


* **T-cells** (identified by high expression of CD3D and CD3E).
* **monocytes** (identified by high expression of LYZ and S100A8).
* **naive B-cells** (identified by high expression of CD79A, CD79B and BLNK).
* **plasma cells** (identified by B-cell and proliferation markers, such as TOP2A or MKI67).
* poor-quality cells (identified by high mitochondrial expression). If a cell has pores in the membrane due to cell lysis, the cytosolic mRNA will leak out of the cell; but if the diameter of mitochondria is higher than the pores, they will get trapped inside the cell.

DZ: SUGCT, CXCR4, AICDA

LZ: CD83, BCL2A1

GC total: MEF2B, BCL6, IRF4

PC: PRDM1, SLAMF7, MZB1, FKBP11


```{r}
canonical_bcell_markers <-c("CD34", "CD38", "CD19")

monocytes_markers<-c("LYZ","S100A8")

naive_markers<-c("CD79A", "CD79B", "BLNK")

bib_Bcell_markers<-c("CD19","CR2","MS4A1","RALGPS2","CD79A")
bib_Tcell_markers<-c("CD3E","CD4","CD8A","FOXP3","IL17A")

markers_bcell<-c("BANK1","ARHGAP24","ADAM28","MARCH1","RAPGEF5","AFF2","RGS13","LPP","IGHG1","IGLC1","SLC8A1","LYZ","PLXDC2","FAM160A1","IGHA1","IGLC2", "SETBP1","ENTPD1","COL19A1","CCSER1")

markers_tcell<-c("INPP4B","FYB1","LEF1","IL7R","IL6ST","CCL5","GNLY","NKG7","DTHD1","RUNX2", "FOXP3","CD8A","IL17A","CD2")

```



## B cells

```{r}
markers_gg(naive_markers)
```


```{r}
markers_gg(bib_Bcell_markers)
 
```


## T-cells

CD8+ T cell markers:"CD3D", "CD8A"
NK cell markers:"GNLY", "NKG7"

```{r}
markers_gg(bib_Tcell_markers)
  
```


## Pauli markers


```{r}
naive_mem_bcell<-c("BANK1", "FCER2")
cd4_tcell<-c("CD3D", "IL7R")
dz_gc_bcell<-c("MKI67", "TOP2A")
lz_gc_bcell<-c("MARCKSL1", "RGS13", "LMO2", "CCDC88A")	
cytotoxic<-c("GNLY", "NKG7", "GZMK", "CD8A")
memory_bcell<-c(	"FCRL4", "FCRL5", "PLAC8", "SOX5")
pc<-c("IGHG1", "IGHA1", "JCHAIN", "XBP1")
myeloid<-c("LYZ", "S100A8")	
poor_q_doublets <-c("FDCSP", "CLU", "CXCL13", "CR2")
doublet_proliferative_tcell<-c("MT2A", "CD3D", "TRAC", "PCNA")
Unk<-c("PTPRCAP", "CD37", "CD74")	
PDC<-c("PTCRA", "LILRA4", "IRF7")
```

```{r}
markers_gg(naive_mem_bcell)
```

```{r}
markers_gg(cd4_tcell)
```

```{r}

markers_gg(dz_gc_bcell)
```

```{r}
markers_gg(lz_gc_bcell)
```
```{r}
markers_gg(cytotoxic)
```

Cytotoxic T cells are effector cells that destroy virus-infected cells, tumor cells, and tissue grafts that exist in the cytosol, or contiguous nuclear compartment. The cells are also known as CD8+ T 

```{r}
markers_gg(pc)
```

```{r}
markers_gg(myeloid)
```

```{r}
markers_gg(poor_q_doublets)
```

```{r}
markers_gg(doublet_proliferative_tcell)
```

 Proliferating cell nuclear antigen (PCNA)
 
```{r}

markers_gg(Unk)
```
```{r}

markers_gg(PDC)

```

```{r}

markers_gg("FOXO1","NFKB1")

```

"NFKB1": CLUSTER 2


```{r}
markers_gg( "MYO1E")
```


## Doublet score


```{r}
FeaturePlot(
    tonsil_wnn_without_doublet,
    features = "doublet_scores",
    reduction = "wnn.umap",
    pt.size = 0.1
  )
```


# CellCycleScoring


```{r}
DimPlot(tonsil_wnn_without_doublet,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T, group.by = "Phase")
```

```{r}
DimPlot(tonsil_wnn_without_doublet,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T)
```
