==> 4.clustering/9-analysing_clusters_bcell.Rmd <==
---
title: "9-Analysing clusters B cells"
author: 
- name: "Maria Lucia Romero Rivero"
  email: MLROMEROR@clinic.cat
date: "`r Sys.Date()`"
output:
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
    fig_captions: yes

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Pre-processing 

## Load package


```{r message=FALSE, warning=FALSE}
library("RColorBrewer")
library(Signac)
library(Seurat)
library(GenomicRanges)
library(future)
#library(SeuratWrappers)
library(harmony)
library(EnsDb.Hsapiens.v86)
library(stringr)
library(dplyr)
library(ggplot2)
library(patchwork)
library(kableExtra)
library(tidyverse)
library(BSgenome.Hsapiens.UCSC.hg38)
set.seed(123)
```


## Parameters

```{r}
# Paths
path_to_obj <- ("~/Documents/multiome_tonsil_Lucia/results/R_objects/13.tonsil_multiome_bcells_without_doublets_normalized.rds")

path_to_markers<-("~/Documents/multiome_tonsil_Lucia/results/tables/tonsil_markers_bcell_075.csv")
path_to_obj_save <- ("~/Documents/multiome_tonsil_Lucia/results/R_objects/13.1.tonsil_multiome_bcells_without_doublets_normalized_peaklink.rds")

```


## Load data

```{r}
tonsil_wnn_bcell <- readRDS(path_to_obj)

tonsil_markers_075<-read_csv(path_to_markers)
```

# UMAP 

```{r}
 DimPlot(
    tonsil_wnn_bcell,
    group.by = "wsnn_res.0.075",
    reduction = "wnn.umap",
    pt.size = 0.1, label = T
  )
```

## Get top n markers of each cluster


**Resolution 0.075**
 
```{r}
top5_tonsil_markers_075<-tonsil_markers_075 %>% group_by(cluster) %>% top_n(n = 5, wt = avg_log2FC)
top7_tonsil_markers_075<-tonsil_markers_075 %>% group_by(cluster) %>% top_n(n = 7, wt = avg_log2FC)
top10_tonsil_markers_075<-tonsil_markers_075 %>% group_by(cluster) %>% top_n(n = 10, wt = avg_log2FC)
```

```{r}
df_top5<-as.data.frame(top5_tonsil_markers_075)
kbl(df_top5,caption = "Table of the top 5 marker of each cluster resolution 0.005") %>%
  kable_paper("striped", full_width = F)
```

```{r}
df_top7<-as.data.frame(top7_tonsil_markers_075)
df_mark<-as.data.frame(tonsil_markers_075)
kbl(df_top7,caption = "Table of the top 7 marker of each cluster resolution 0.005") %>%
  kable_paper("striped", full_width = F)
```

### Dynamic table top 7

```{r}
#install.packages("htmlwidgets", type = "binary")
#install.packages("DT", type = "binary")

DT::datatable(df_top7)

```


### Dynamic table of all markers

```{r}
DT::datatable(df_mark)

```


### Dotplot

```{r}
markerGenes <- unique(top10_tonsil_markers_075$gene)

Idents(tonsil_wnn_bcell) <- "wsnn_res.0.075"

dot.10 <- DotPlot(tonsil_wnn_bcell, features = unique(top10_tonsil_markers_075$gene),cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) +ggtitle("res 0.1 top 10 of each cluster")

dot.5 <- DotPlot(tonsil_wnn_bcell, features = unique(top5_tonsil_markers_075$gene),cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("res 0.1 top 5 of each cluster")


dot.5.nscale <- DotPlot(tonsil_wnn_bcell, features = top5_tonsil_markers_075$gene,cols = c("lightgrey","blue"), cluster.idents = T) + theme(axis.text.x = element_text( size = 10, vjust = 0.8, hjust = 0.8)) + ggtitle("res 0.1 top 5 of each cluster no scale")

```

```{r}

dot.10 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 5))
dot.5 +
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 7))
dot.5.nscale+
  coord_flip() +
  theme(axis.title = element_blank(), axis.text.y = element_text(size = 7))

```


# Markers exploration

```{r}

function(tonsil_markers_075,n){
  top_tonsil_markers_075<-tonsil_markers_075 %>% group_by(cluster) %>% top_n(n = n, wt = avg_log2FC)

}
top7mark_cluster0<-top7_tonsil_markers_075[["gene"]][1:7]
top7mark_cluster1<-top7_tonsil_markers_075[["gene"]][8:14]
top7mark_cluster2<-top7_tonsil_markers_075[["gene"]][15:21]
top7mark_cluster3<-top7_tonsil_markers_075[["gene"]][22:28]
top7mark_cluster4<-top7_tonsil_markers_075[["gene"]][29:35]
top7mark_cluster5<-top7_tonsil_markers_075[["gene"]][36:42]

```


```{r}
markers_gg <- function(x){purrr::map(x, function(x) {
  
  p <- FeaturePlot(
    tonsil_wnn_bcell,
    features = x,
    reduction = "wnn.umap",
    pt.size = 0.1
  )
  p
})}
```


## Main markers 

```{r}
m<-c("PRDM1","XBP1","IRF4","MEF2B","BCL6")

DZ<-c("SUGCT", "CXCR4", "AICDA")

LZ<- c("CD83","BCL2A1")

GC<- c("MEF2B", "BCL6","IRF4")

PC<- c("PRDM1","SLAMF7", "MZB1", "FKBP11")

```

### Dark Zone

```{r}
markers_gg(DZ)

```

### Light Zone

```{r}
markers_gg(LZ)

```

### Germinal Center


```{r}
markers_gg(GC)

```

### Plasma cell


```{r}
markers_gg(PC)

```
## Pauli markers


```{r}
naive_mem_bcell<-c("BANK1", "FCER2")
cd4_tcell<-c("CD3D", "IL7R")
dz_gc_bcell<-c("MKI67", "TOP2A")
lz_gc_bcell<-c("MARCKSL1", "RGS13", "LMO2", "CCDC88A")	
cytotoxic<-c("GNLY", "NKG7", "GZMK", "CD8A")
memory_bcell<-c(	"FCRL4", "FCRL5", "PLAC8", "SOX5")
pc<-c("IGHG1", "IGHA1", "JCHAIN", "XBP1")
myeloid<-c("LYZ", "S100A8")	
poor_q_doublets <-c("FDCSP", "CLU", "CXCL13", "CR2")
doublet_proliferative_tcell<-c("MT2A", "CD3D", "TRAC", "PCNA")
Unk<-c("PTPRCAP", "CD37", "CD74")	
PDC<-c("PTCRA", "LILRA4", "IRF7")
```

```{r}
markers_gg(c("TCL1A","IGHD","FCER2","IGHM","IL4R"))

```

```{r}
markers_gg(naive_mem_bcell)
```
```{r}
markers_gg(memory_bcell)
```


```{r}

markers_gg(dz_gc_bcell)
```

```{r}
markers_gg(lz_gc_bcell)
```


```{r}
markers_gg(pc)
```


```{r}
markers_gg(poor_q_doublets)
```



 Proliferating cell nuclear antigen (PCNA)
 
```{r}

markers_gg(Unk)
```

```{r}

markers_gg(PDC)

```


```{r}
markers_gg( "MYO1E")
```




## Top 7 of each cluster 

### Cluster 0

```{r}
markers_gg("COL4A3")
```

FCRL4 is an immunoregulatory receptor that belongs to the Fc receptor-like (FCRL) family. In healthy individuals, FCRL4 is specifically expressed by memory B cells (MBCs) localized in sub-epithelial regions of lymphoid tissues. Expansion of FCRL4+ B cells has been observed in blood and other tissues in various infectious and autoimmune disorders

```{r}
markers_gg(top7mark_cluster0)
```



```{r}
naive_markers<-c("CD79A", "CD79B", "BLNK")
memory_markers<-c("CD27")

```

```{r}
markers_gg(naive_markers)
```

```{r}
markers_gg(memory_markers)

```

```{r}
markers_gg(c("MS4A1","NT5E"))
```
MS4A1:NAIVE-MEMORY B CELL
NT5E: NAIVE B CELL 

### Cluster 1: 


```{r}
markers_gg(top7mark_cluster1)
```


### Cluster 2: 


```{r}
markers_gg(top7mark_cluster2)
```



### Cluster 3:

```{r}
markers_gg(top7mark_cluster3)
```
INPP4B: Immune cell enhanced (memory CD4 T-cell)



### Cluster 4

AOAH: NK
GNLY:NK


```{r}
markers_gg(top7mark_cluster4)
```


### Cluster 5

```{r}
markers_gg(top7mark_cluster5)
```


## Markers FOXO1, NFKB1

```{r}
 p <- FeaturePlot(
    tonsil_wnn_bcell,
    features = c("NFKB1","FOXO1"),
    reduction = "wnn.umap",
    pt.size = 0.1
  )
p
```

# Mitocondrial 

```{r}
FeaturePlot(
    tonsil_wnn_bcell,
    features = "percent.mt",
    reduction = "wnn.umap",
    pt.size = 0.1
  )
```

```{r}
VlnPlot(tonsil_wnn_bcell, features = "percent.mt", group.by = "wsnn_res.0.075", pt.size=0)
```

```{r}
VlnPlot(tonsil_wnn_bcell, features = "percent_ribo", group.by = "wsnn_res.0.075", pt.size=0)
```


```{r}
weight.vp<- VlnPlot(tonsil_wnn_bcell, features = "RNA.weight", group.by = "wsnn_res.0.075",pt.size = 0)
p.weight.vp<- VlnPlot(tonsil_wnn_bcell, features = "peaks.weight", group.by = "wsnn_res.0.075",pt.size = 0)

weight.vp + ggtitle("RNA modality weight")
p.weight.vp+ ggtitle("ATAC modality weight")
```


# Number of cell in each cluster

```{r}
Idents(tonsil_wnn_bcell) <- "wsnn_res.0.075"

cell.num <- table(Idents(tonsil_wnn_bcell))
cell.num
```

# Rename clusters

```{r}
new.cluster.ids <- c("MBC", "NBC","GC/DZ", "GC/LZ", "GC/PC", "PC")
names(new.cluster.ids) <- levels(tonsil_wnn_bcell)
tonsil_wnn_bcell <- RenameIdents(tonsil_wnn_bcell, new.cluster.ids)
DimPlot(tonsil_wnn_bcell, reduction = "wnn.umap", label = TRUE, pt.size = 0.5) 
```



# Bibliography  markers

**MARKERS**

Immature B cells express CD19, CD 20, CD34, CD38, and CD45R,
T-cell receptor/CD3 complex (TCR/CD3 complex) 


* **T-cells** (identified by high expression of CD3D and CD3E).
* **monocytes** (identified by high expression of LYZ and S100A8).
* **naive B-cells** (identified by high expression of CD79A, CD79B and BLNK).
* **plasma cells** (identified by B-cell and proliferation markers, such as TOP2A or MKI67).
* poor-quality cells (identified by high mitochondrial expression). If a cell has pores in the membrane due to cell lysis, the cytosolic mRNA will leak out of the cell; but if the diameter of mitochondria is higher than the pores, they will get trapped inside the cell.



```{r}
canonical_bcell_markers <-c("CD34", "CD38", "CD19")

monocytes_markers<-c("LYZ","S100A8")

naive_markers<-c("CD79A", "CD79B", "BLNK")

bib_Bcell_markers<-c("CD19","CR2","MS4A1","RALGPS2","CD79A")
bib_Tcell_markers<-c("CD3E","CD4","CD8A","FOXP3","IL17A")


```



## B cells

```{r}
markers_gg(naive_markers)
```


```{r}
markers_gg(bib_Bcell_markers)
 
```


## T-cells

CD8+ T cell markers:"CD3D", "CD8A"
NK cell markers:"GNLY", "NKG7"

```{r}
markers_gg(bib_Tcell_markers)
  
```

```{r}
DimPlot(tonsil_wnn_bcell,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T, group.by = "age_group")
```


# CellCycleScoring


```{r}
DimPlot(tonsil_wnn_bcell,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T, group.by = "Phase")
```
"
```{r}
DimPlot(tonsil_wnn_bcell,
    reduction = "wnn.umap",
    pt.size = 0.1, label = T)
```

```{r}
FeaturePlot(tonsil_wnn_bcell,
            reduction = "wnn.umap",
            features = "doublet_scores")
```

#Link genes

```{r}
DefaultAssay(tonsil_wnn_bcell) <- "peaks"

# first compute the GC content for each peak
tonsil_wnn_bcell <- RegionStats(tonsil_wnn_bcell, genome = BSgenome.Hsapiens.UCSC.hg38)

# link peaks to genes
tonsil_wnn_bcell <- LinkPeaks(
  object = tonsil_wnn_bcell,
  peak.assay = "peaks",
  expression.assay = "RNA"
)
```
# Rename clusters

```{r}
new.cluster.ids <- c("MBC", "NBC","GC/DZ", "GC/LZ", "GC/PC", "PC")
names(new.cluster.ids) <- levels(tonsil_wnn_bcell)
tonsil_wnn_bcell <- RenameIdents(tonsil_wnn_bcell, new.cluster.ids)
DimPlot(tonsil_wnn_bcell, reduction = "wnn.umap", label = TRUE, pt.size = 0.5) 
```


```{r}
idents.plot <- c("MBC", "NBC","GC/DZ", "GC/LZ", "GC/PC", "PC")

p1 <- CoveragePlot(
  object = tonsil_wnn_bcell,
  region = "ANK3",
  features = "ANK3",
  expression.assay = "RNA",
  idents = idents.plot,
  extend.upstream = 1000,
  extend.downstream = 10000
)

p2 <- CoveragePlot(
  object = tonsil_wnn_bcell,
  region = "COL19A1",
  features = "COL19A1",
  expression.assay = "RNA",
  idents = idents.plot,
  extend.upstream = 8000,
  extend.downstream = 10000
)
p1
p2
#patchwork::wrap_plots(p1, p2, ncol = 1)
```


# 3D UMAP
 This is a manual method of extracting embeddings and is not needed


```{r}
umap_plot3d(
  tonsil_wnn_bcell,
  coloring = "clusters",
  reduction = "wnn.UMAP"
)
```


```{r eval=FALSE, include=FALSE}
yourseuratobject <- RunUMAP(tonsil_wnn_bcell,
                            dims = ,
                            n.components = 3L)

# Extract tSNE information from Seurat Object
umap_1 <- yourseuratobject[["umap"]]@cell.embeddings[,1]
umap_2 <- yourseuratobject[["umap"]]@cell.embeddings[,2]
umap_3 <- yourseuratobject[["umap"]]@cell.embeddings[,3]

# as pointed out by user @sdinardo on 01142022 thank you! 
# Extract UMAP information from Seurat Object
UMAP_1 <- tonsil_wnn_bcell[["wnn.umap"]]@cell.embeddings[,1]
UMAP_2 <- tonsil_wnn_bcell[["wnn.umap"]]@cell.embeddings[,2]
UMAP_3 <- tonsil_wnn_bcell[["wnn.umap"]]@cell.embeddings[,3]

# Visualize what headings are called so that you can extract them to form a dataframe
Embeddings(object = tonsil_wnn_bcell, reduction = "wnn.umap")

# Prepare a dataframe for cell plotting
plot.data <- FetchData(object = tonsil_wnn_bcell, vars = c("wnnUMAP_1", "wnnUMAP_2", "wnnUMAP_3", "seurat_clusters"))

# Make a column of row name identities (these will be your cell/barcode names)
plot.data$label <- paste(rownames(plot.data))

# Plot your data, in this example my Seurat object had 21 clusters (0-20)
fig <- plot_ly(data = plot.data, 
        x = ~UMAP_1, y = ~UMAP_2, z = ~UMAP_3, 
        color = ~seurat_clusters, 
        colors = c("lightseagreen",
                   "gray50",
                   "darkgreen",
                   "red4",
                   "red",
                   "turquoise4",
                   "black",
                   "yellow4",
                   "royalblue1",
                   "lightcyan3",
                   "peachpuff3",
                   "khaki3",
                   "gray20",
                   "orange2",
                   "royalblue4",
                   "yellow3",
                   "gray80",
                   "darkorchid1",
                   "lawngreen",
                   "plum2",
                   "darkmagenta"),
        type = "scatter3d", 
        mode = "markers", 
        marker = list(size = 5, width=2), # controls size of points
        text=~label, #This is that extra column we made earlier for which we will use for cell ID
        hoverinfo="text") #When you visualize your plotly object, hovering your mouse pointer over a point shows cell names
 

```
```{r}
c<-for (i in seq_along(1:5)) {
  n<-paste0("top7mark_cluster",i,"") 
  return(n)
}

print(c)
```


```{r}
saveRDS(tonsil_wnn_bcell, path_to_obj_save)

```
